all done

+++ libft/ft_atoi.c
@@ -6,17 +6,17 @@
 /*   By: sharrach <sharrach@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/03 15:34:10 by sharrach          #+#    #+#             */
-/*   Updated: 2023/01/24 19:50:31 by sharrach         ###   ########.fr       */
+/*   Updated: 2023/01/25 11:01:05 by sharrach         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-long long	ft_atoi(const char *str)
+int	ft_atoi(const char *str)
 {
 	int			i;
 	int			sign;
-	long long	nb;
+	int			nb;
 
 	i = 0;
 	sign = 1;
